RTC and Time updates
The library used for interacting with the RTC was changed from RTClib to DS3232RTC. Besides integrating nicer with the ardruino's system clock, this library also makes it easier to use the RTC's alarm and temp features.  The time library was also updated to PJRC's Time library. These updates also fix the bugs caused by changing to the teensy board.